Add files from earlier Python project
Significantly refactored code outside of the history for this project.

Added tools for extracting paths from SVG files and saving the path's
coordinates to CSV files to be used in developing this project.

Added tool for generating points within a path and exporting those
points to a CSV file, again to be used in developing this project.