Use drag/dropped .csv file for table
Before, I chose the `.csv` file for the table using the HTML file
selection dialog.  So far the only way I've figured out how to read
a file in PaperScript is by dragging and dropping a file into the
document, so for now I'll populate the data table with the same
drag-and-dropped file and remove the ability to chose a file with the
HTML file selection dialog.

Also, I renamed progress.js to data-table.js.