Set y translation according to the first file
I had each data set being flipped vertically over the x-axis, and then
slid up to be just barely above the x-axis.  But that's wrong.  That
way, what were originally the tops of all the datasets end up at the
same level, but different data sets have different max y values, so
that's wrong.  What should happen is each data set should be flipped
over the x-axis, and then be shifted up some CONSTANT amount, rather
than a different amount for each data set, in order for the data points
to maintain their positions relative to each other.

For now, I'm translating the first data set like I did originally;
putting what was originally the y max at y=0, but then I shift any
subsequent datasets up by that SAME amount.

If a subsequent data set had a y max greater than that of the first data
set, that would mean part of it would lie below the x-axis and not be
visible on the screen.  But that's okay for this application, because
the first data set should be the outline of a raster, and subsequent
data sets should be random points WITHIN the outline (i.e. the
subsequent data set's y max has to be less that the first data set's).