Fix y orientation
Now, the data is used in the same orientation as it is in the `.svg`
file, and only flipped when being plotted with matplotlib.

Before, I wasn't thinking carefully and thought the data imported from
the `.svg` file was oriented with y increasing up, which is the wrong
convention for computer graphics, and so I inverted it to match the
convention.  Turns out that was wrong, and the `.svg` data already
conform to y-increasing-down.

The only place I need to invert the direction is when using matplotlib,
which uses the y-increasing-up convention of mathematics and science.