Implement drawing a polyline
Click to add a node to the polyline.
Click an existing node to toggle whether it is selected.

Added SCSS to the project.

Tried to systematize event handling:
- Start names of callbacks with "do", so the code reads as "target: on
event, do action".
  e.g.
       drawing.on("mousedown", doStartOutlining);
       "drawing: on mousedown, do start outlining"
- Don't use Event.stopPropagation; it's the callback's job to decide
whether or not to execute based on the event's history: which nodes it
has already bubbled through, what the event's original target was, etc.