Start a refactoring project.  Fist two files.
The idea is to store all the D3 stuff in one object, all the Paper.js
stuff in another object, and all the objective/agnostic/actually
important data that defines the project in another object.

I'm thinking this will make it easier to separate the rendering stuff
from the calculating stuff later.

I might also convert the whole `app` function into a `LowPolyApp` class.

I really should read up on how to build a web app, but that'll come
later.