Refactor to use underlying ES prototypical OO mechanism
In the version with the new `class` syntax, I still have to
`Object.assign` the methods that are defined in other files.  It seems
kind of incongruous to use those together, so I made this version which
uses the underlying stuff the whole way through, which seems more
self-consistent.

The new syntax could be improved with something similar to the the
object deconstructing syntax, and my need to use `Object.assign` would
be eliminated:
    function foo() {...}
    function bar() {...}
    function baz() {...}

    class Cricket {
         constructor() {...}

         // This is the currently allowed method declaration syntax
	 chicken() {...}

	 // This is the proposed improvement:
	 //   You could just list an existing function like you do with
	 //   object deconstructing i.e.
	 //     const obj = { lemon: "quatrain", pigeons: 14 };
	 //     const { lemon, pigeons: dessert };
	 //
	 //     >>> lemon
	 //     "quatrain"
	 //     >>> dessert
	 //     14
	 foo

	 // Proposed renaming like object deconstructing
         //   Putting `bar` and `newNameForBar` in the other order is
	 //   intuitive IMO, but doing it this way keeps it in parallel
	 //   with object deconstruction.
	 bar: newNameForBar

	 // Defining a method to execute an existing function works
	 // currently, but it is weird and ugly and I don't like it.
	 baz(...args) {
	     return baz(...args);
	 }
    }