Drag a node to change all polygons that include that node
Major success.  Restructured the way data is stored.  Implemented some
new utility functions.

Now, each node is represented by a single object (an array containing
its x and y coordinates).  Anything that includes that node (a polygon,
an edges) doe so by referencing that object.  Thus, when a node is
updated, everything that includes that node can detect the change.

Before, each polygon had its own copy of each of the nodes that made up
its vertices, so dragging a node only changed one polygon.

This could be implemented more efficiently, but as it is I'm not having
any performance issues.

Right now, whenever a node is dragged, ALL polygons are updated.  Only
the polygons that include the point that was dragged NEED to be updated,
though.

Also, the way I ensure that there's only one object representing each
node is a little shakey.  It goes like this:
1. I generate a Voronoi diagram from random points.
2. I get a list of polygons formed by the Voronoi diagram (at this
   point, shared points are represented by *different* objects, so an
   update to one doesn't affect the others).
3. I trim off parts of polygons that lie outside the outline of the
   raster.  That process generates a whole new set of polygons.
***4. I go through each polygon, and add each of its points to a global
   list of nodes.  For each point, I check if there's already an object
   in the global list with the same coordinates.  If there is, I get rid
   of the object in the polygon's list and replace it with a reference
   to the one in the global list. (I can't use a `Set` object because
   the nodes are mutable.)

***THIS IS THE IFFY PART
This relies on the assumption that both (1) the function that trims the
polygons, and (2) the function that generates polygons from the Voronoi
diagram, use EXACTLY the same values for points shared between two or
more polygons.  I was surprised to find that this works, especially
since the coordinates have A LOT of digits after the decimal.

A quick fix to make this more robust, if necessary, would be to
introduce some fudge factor/epsilon to the function that compares
points.

A more principled fix would be to delve into the implementation of both
of those functions to make sure they use the same objects for identical
points to begin with.