Make kill action for ending stage. Unplug UI listeners on kill.
The sites don't remove their listeners on kill yet; if you click on them
after `kill` has been dispached, they'll still delete themselves.  The
way I have siteElement implemented right now is kind of hacky, and I'm
going to redo it with react, so I'll make them respond to `kill` then.
It would be inelegant at best if I tried to implement it now, and it's
just going to get redone anyway so...