Implement Redux, React-Redux for sites
State for the sites phase is now in a Redux store.

So far, only the sites interact with the Redux mechanism.

A _presentation component_ `Sites` renders the sites.  It recieves the
current sites and an event handler from its parent, a _container
component_ `SitesContainer`, which is created programmatically by the
React-Redux function `connect`.  It takes the Redux store as a property,
and uses it to extract the list of sites from the state, which is passes
as a property to `Sites`, and wraps `dispatch` as a function that it
passes, also as a prop, to `Sites` to be used as an event handler.
`SitesContainer` is not explicitly passed the store; it is wrapped in a
component `Provider` from React-Redux which passes the store as a prop
to all its descendants.

Clicking on sites deletes them by dispatching actions to the Redux store.

Other actions can be handled by the React reducer, but the other UI
React components (`Outline`, `NumPicker`, `EndStage`) don't currently
interact with the Redux mechanism.  They generate the right actions for
the right events, but they don't dispatched them to the Redux store yet;
they just log them to the console.