Render raster at full size. Make SVG dimensions equal raster dimensions.
The app reads the raster's dimensions asyncronously by loading the
raster into an `HTMLImageElement` (that we only use to get the raster's
dimensions).  Then it saves the dimensions to the store, even though the
storing the dimensions is redundant becuase they are derivable from from
the raster, which is also stored.  But this way we only have to do the
blocking/asynchronous operation once.