Click to add outline nodes. Move selection listners to OutlineToolUI.
\#There're some bugs to fix and clean up to do:

- BUG: I don't have event propogation worked out right.  Listeners are
currently being executed on both the `patch` and the `workspace` for the
same event.  Using `stopPropagation` didn't work as expected; looked
like the listener for `workspace` was being triggered first, which is
the opposite of what I expected, since the default behavior is for
events to bubble from children to parents; `patch` is a descendant of
`workspace`, so I expected it's listener to be triggered first, but that
doesn't seem to be what's happening.
Right now I don't have `stopPropagation` running, so both listeners are
executing when the `patch` is clicked, which happens to result in the
desired behavior for selecting or deselecting a patch, as bubbling seems
to be working backward, with the listener on `workspace` executing first
(clearing the selection), and then the listener on `patch` executing
(selecting the patch).  Adding nodes has the wrong behavior though;
clicking on the `patch`'s outline adds two nodes: one for the listener
on the `patch`, and one for the listener on the `workspace`.

- CLEAN UP: In the previous commit, instead of installing the selection
listeners in `OutlineToolUI`, they were installed directly on
`workspace` when it is first rendered.  I haven't changed
`DefaultToolUI` yet, so it still assumes the old behavior.  It installs
a selection listener on the `patch`, assuming you can clear the
selection by clicking on blank space, but this is no longer the case.