Add and remove listeners when tool/subtool changes
Extract outline tool's rendering of patch to a separate component.

Add and remove listeners at appropriate lifecycle stages so they don't
linger when the tool or subtool changes.  Needing to do this is the cost
of imperatively installing the listeners, rather than declaratively
installing them when a component is first created, as works more
naturally with React's one way data flow.  I'm doing it this way because
it seems like it's the toolUI's job to decide how the `svg.workspace`
element should handle events (in this case nothing/clearing the
selection/adding a node for modes null/select/edit).  I could have
`svg.workspace` respond to changes to the tool and store all of its
possible handlers itself, but that would be mixing concerns and could
grow to be messy in my view.  The way I have it, the tool decides what
`svg.workspace`'s handler should be and then it installs it (ie
imperatively).  One way to keep everything declarative and keep data
flowing one way would be to decide what handler `svg.workspace` should
have within the tool, then dispatch an action that stores that
function in state, and then have `svg.workspace` read that function from
state and rerender.  I can't do that now, because I deep copy state by
turning the old state into a JSON string and then parsing it, and I
believe JSON can't hold functions.  I'm thinking switching to
immutable.js might allow me to store functions in state (and would also
automatically deep copy things), so I might switch to that eventually,
but I'd rather move forward with writing this app rather than stop
progress for a while to learn immutable.js.

I may also want to look into JQuery's event handling stuff, which I
think might enable me to attach data to an event as it bubbles, which
could help eliminate the need for using stopPropagation to stop events
from triggering both `svg.workspace` and `.patch` listeners, which would
be nice because `stopPropagation`ing is a rather harsh way to achieve
that goal.  But same deal as with immutable.js; don't want to slow down
to learn JQuery at the moment.